Fix MXINT8 RTL tests and accuracy tests. Implement logic for vector unit matrix transpose.
Former-commit-id: bf2833761fe5ae870c598abe47bc8db69d2cfa67

+++ test/common/VerificationTypes.h
@@ -88,13 +88,13 @@ inline std::vector<int> get_shape(const codegen::Tensor& tensor) {
   return {repeated_field.begin(), repeated_field.end()};
 }
 
-inline int get_size(const std::vector<int>& shape) {
-  int size = 1;
+inline long get_size(const std::vector<int>& shape) {
+  long size = 1;
   for (const auto& dim : shape) size *= dim;
   return size;
 }
 
-inline int get_size(const codegen::Tensor& tensor) {
+inline long get_size(const codegen::Tensor& tensor) {
   const auto shape = get_shape(tensor);
   return get_size(shape);
 }
